Clarify relationship between env and ini mode
On [this question on Stack Overflow](https://stackoverflow.com/q/65107145/157957), a user was confused that setting `XDEBUG_MODE` didn't change the value of the `xdebug.mode` ini setting. As far as I can see, that's intentional, and the actual setting was taking effect as expected, but the wording here was a bit ambiguous.